- revamp authentication provider interface for java
